backend/requirements.txt
Django==3.1
django-bootstrap-form==3.4
djangorestframework==3.11.1
Pillow==7.2.0